pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cape"
version = "0.1.0"
description = "Capability Package - Model-agnostic capability abstraction for AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "your@email.com" }
]
keywords = ["capability", "agent", "ai", "orchestration", "langchain", "openai", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
langchain = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.18.0",
]
embeddings = [
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.0",
]
all = [
    "cape[langchain,openai,anthropic,embeddings]",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["cape"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
